Optimization : no GUI latency (add stack of Tasks in a Thread database)
